refactor: move design system to root directory
