+++ composeApp/src/commonMain/kotlin/login/LoginScreen.kt
@@ -18,15 +18,15 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
-import designsystem.colors.ButtonColor
-import designsystem.login.CredentialInput
-import designsystem.login.FacebookSignUpHyperlink
-import designsystem.login.ForgotPasswordHyperlink
-import designsystem.login.InstagramTextLogo
-import designsystem.login.InstagramSignUpHyperlink
-import designsystem.login.LoginButton
-import designsystem.login.LoginDivider
-import designsystem.login.LoginFooter
+import colors.ButtonColor
+import components.buttons.LoginButton
+import components.inputs.CredentialInput
+import components.misc.LoginDivider
+import components.misc.LoginFooter
+import components.texts.FacebookSignUpHyperlink
+import components.texts.ForgotPasswordHyperlink
+import components.texts.InstagramSignUpHyperlink
+import components.texts.InstagramTextLogo
 import dev.icerock.moko.resources.compose.painterResource
 import dev.icerock.moko.resources.compose.stringResource
 import instaclone.resources.MR
