+++ design-system/src/commonMain/kotlin/colors/TextColor.kt
@@ -1,4 +1,4 @@
-package designsystem.colors
+package colors
 
 import androidx.compose.ui.graphics.Color
