+++ design-system/src/commonMain/kotlin/components/inputs/CredentialInput.kt
@@ -1,6 +1,6 @@
-package designsystem.login
+package components.inputs
 
-import designsystem.colors.ButtonColor
+import colors.ButtonColor
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -20,7 +20,7 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
-import designsystem.colors.TextColor
+import colors.TextColor
 import dev.icerock.moko.resources.ImageResource
 import dev.icerock.moko.resources.compose.painterResource
 import instaclone.resources.MR
