+++ design-system/src/commonMain/kotlin/components/misc/OrDivider.kt
@@ -1,4 +1,4 @@
-package designsystem.login
+package components.misc
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
@@ -10,7 +10,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import designsystem.colors.TextColor
+import colors.TextColor
 
 @Composable
 fun LoginDivider() {
